pyproject.toml
[project]
name = "mp3-uploader"
version = "0.1.0"
description = ""
authors = [
    {name = "anton",email = "antonkuzmn1@gmail.com"}
]
readme = "README.md"
requires-python = "==3.13.9"
dependencies = [
    "poetry-core (>=2.2.1,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "aiogram (>=3.23.0,<4.0.0)",
    "yt-dlp (>=2025.12.8,<2026.0.0)",
    "fastapi (>=0.127.1,<0.128.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "python-multipart (>=0.0.21,<0.0.22)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false